Overhaul of existing imports system for index use
Replace pre-imports with simple import-item gen. stage

- pre-imports isn't going to work out with indexing; will write about my tests with this
- removing the old logic, and now what happens in the pre-imports stage is a simple mapping from Web Ext. history/bookmarks API to our import item list in local storage
- a lot of other stuff has been refactored out of imports-prep as it's still needed elsewhere (mostly transformations that need to get moved anyway)

Increase data stored in import items from browser APIs

- as there are no stubs, we need to store the browser ID + timestamp in the import items now (not really a bad thing)
- also turn off imports concurrency (as the indexing method is not concurrent safe)

Update history import item processing logic for new flow

- instead of updating an existing pouch doc with fetched content, this now constructs the new doc and inserts
 - it also calls the indexing method (which, if you try, slows things down by a lot)
- TODO: adding in visits gathering
- TODO: differentiate bookmarks logic (in old imports, they were exactly the same at this stage)

Implement visits in new import-item-processor

- also reverted back to using conncurrent index add :\ it just seems to slow to sit there and wait for up to a minute per item; now imports still feels fast but the indexing is still slow happening in the background
- need a way to check index ready state

Implement bookmarks in new import-item-processor

- similar to the old logic, but now doing it at later stage (old full-imports stage)
