+++ src/background.js
@@ -9,7 +9,7 @@ import { installTimeStorageKey } from 'src/imports/background'
 import convertOldData from 'src/util/old-data-converter'
 import fetchPageData from 'src/page-analysis/background/fetch-page-data'
 import { findPagesByUrl } from 'src/search/find-pages'
-import { transformToBookmarkDoc } from 'src/imports/background/imports-preparation'
+import { transformToBookmarkDoc } from 'src/imports'
 import { generatePageDocId } from 'src/page-storage'
 import { FREEZE_DRY_BOOKMARKS_KEY } from 'src/options/preferences/constants'
 import * as index from 'src/search/search-index'
